requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.1
joblib==1.3.2
requests==2.31.0
tqdm==4.65.0
imbalanced-learn==0.10.1
sgp4==2.22
seaborn
matplotlib.pyplot
